+++ spec/at_coder_friends/cxx_generator_spec.rb
@@ -7,52 +7,25 @@ RSpec.describe AtCoderFriends::Generator::CxxBuiltin do
   let(:cfg) { nil }
 
   describe '#select_template' do
-    subject { generator.select_template(interactive) }
+    subject { generator.select_template }
 
     context 'with default configuration' do
-      context 'for interactive problems' do
-        let(:interactive) { true }
-
-        it 'returns template file name' do
-          expect(subject).to eq(
-            File.join(TMPL_DIR, 'cxx_builtin_interactive.cxx')
-          )
-        end
-      end
-
-      context 'for other problems' do
-        let(:interactive) { false }
-
-        it 'returns template file name' do
-          expect(subject).to eq(
-            File.join(TMPL_DIR, 'cxx_builtin_default.cxx')
-          )
-        end
+      it 'returns template file name' do
+        expect(subject).to eq(
+          File.join(TMPL_DIR, 'cxx_builtin.cxx.erb')
+        )
       end
     end
 
     context 'with custom configuration' do
       let(:cfg) do
         {
-          'default_template' => 'customized_default.cxx',
-          'interactive_template' => 'customized_interactive.cxx'
+          'default_template' => 'customized_default.cxx'
         }
       end
 
-      context 'for interactive problems' do
-        let(:interactive) { true }
-
-        it 'returns template file name' do
-          expect(subject).to eq('customized_interactive.cxx')
-        end
-      end
-
-      context 'for other problems' do
-        let(:interactive) { false }
-
-        it 'returns template file name' do
-          expect(subject).to eq('customized_default.cxx')
-        end
+      it 'returns template file name' do
+        expect(subject).to eq('customized_default.cxx')
       end
     end
   end
@@ -245,7 +218,7 @@ RSpec.describe AtCoderFriends::Generator::CxxBuiltin do
     end
 
     context 'for a jagged array of numbers' do
-      let(:container) { :vmatrix }
+      let(:container) { :varray_matrix }
       let(:item) { :number }
       let(:names) { %w[K A] }
       let(:size) { %w[N K_N] }
@@ -260,7 +233,7 @@ RSpec.describe AtCoderFriends::Generator::CxxBuiltin do
     end
 
     context 'for a jagged array of characters' do
-      let(:container) { :vmatrix }
+      let(:container) { :varray_matrix }
       let(:item) { :char }
       let(:names) { %w[K p] }
       let(:size) { %w[Q 26] }
@@ -392,7 +365,7 @@ RSpec.describe AtCoderFriends::Generator::CxxBuiltin do
     end
 
     context 'for a jagged array of numbers' do
-      let(:container) { :vmatrix }
+      let(:container) { :varray_matrix }
       let(:item) { :number }
       let(:names) { %w[K A] }
       let(:size) { %w[N K_N] }
@@ -409,7 +382,7 @@ RSpec.describe AtCoderFriends::Generator::CxxBuiltin do
     end
 
     context 'for a jagged array of characters' do
-      let(:container) { :vmatrix }
+      let(:container) { :varray_matrix }
       let(:item) { :char }
       let(:names) { %w[K p] }
       let(:size) { %w[Q 26] }
@@ -460,7 +433,7 @@ RSpec.describe AtCoderFriends::Generator::CxxBuiltin do
     subject { generator.generate(pbm) }
     let(:pbm) do
       AtCoderFriends::Problem.new('A') do |pbm|
-        pbm.formats_raw = formats
+        pbm.formats_src = formats
         pbm.constants = constants
         pbm.options.interactive = interactive
       end
@@ -474,7 +447,9 @@ RSpec.describe AtCoderFriends::Generator::CxxBuiltin do
       end
       let(:formats) do
         [
-          AtCoderFriends::Problem::InputFormat.new(:single, :number, %w[N M]),
+          AtCoderFriends::Problem::InputFormat.new(
+            :single, :number, %w[N M], []
+          ),
           AtCoderFriends::Problem::InputFormat.new(
             :varray, :number, %w[A B C T], %w[M]
           )
@@ -543,7 +518,9 @@ RSpec.describe AtCoderFriends::Generator::CxxBuiltin do
       end
       let(:formats) do
         [
-          AtCoderFriends::Problem::InputFormat.new(:single, :number, %w[N Q])
+          AtCoderFriends::Problem::InputFormat.new(
+            :single, :number, %w[N Q], []
+          )
         ]
       end
       let(:constants) do