+++ spec/at_coder_friends/ruby_generator_spec.rb
@@ -7,27 +7,13 @@ RSpec.describe AtCoderFriends::Generator::RubyBuiltin do
   let(:cfg) { nil }
 
   describe '#select_template' do
-    subject { generator.select_template(interactive) }
+    subject { generator.select_template }
 
     context 'with default configuration' do
-      context 'for interactive problems' do
-        let(:interactive) { true }
-
-        it 'returns template file name' do
-          expect(subject).to eq(
-            File.join(TMPL_DIR, 'ruby_builtin_interactive.rb')
-          )
-        end
-      end
-
-      context 'for other problems' do
-        let(:interactive) { false }
-
-        it 'returns template file name' do
-          expect(subject).to eq(
-            File.join(TMPL_DIR, 'ruby_builtin_default.rb')
-          )
-        end
+      it 'returns template file name' do
+        expect(subject).to eq(
+          File.join(TMPL_DIR, 'ruby_builtin.rb.erb')
+        )
       end
     end
 
@@ -35,24 +21,10 @@ RSpec.describe AtCoderFriends::Generator::RubyBuiltin do
       let(:cfg) do
         {
           'default_template' => 'customized_default.rb',
-          'interactive_template' => 'customized_interactive.rb'
         }
       end
-
-      context 'for interactive problems' do
-        let(:interactive) { true }
-
-        it 'returns template file name' do
-          expect(subject).to eq('customized_interactive.rb')
-        end
-      end
-
-      context 'for other problems' do
-        let(:interactive) { false }
-
-        it 'returns template file name' do
-          expect(subject).to eq('customized_default.rb')
-        end
+      it 'returns template file name' do
+        expect(subject).to eq('customized_default.rb')
       end
     end
   end
@@ -223,7 +195,7 @@ RSpec.describe AtCoderFriends::Generator::RubyBuiltin do
     end
 
     context 'for a jagged array of numbers' do
-      let(:container) { :vmatrix }
+      let(:container) { :varray_matrix }
       let(:item) { :number }
       let(:names) { %w[K A] }
       let(:size) { %w[N K_N] }
@@ -241,7 +213,7 @@ RSpec.describe AtCoderFriends::Generator::RubyBuiltin do
     end
 
     context 'for a jagged array of characters' do
-      let(:container) { :vmatrix }
+      let(:container) { :varray_matrix }
       let(:item) { :char }
       let(:names) { %w[K p] }
       let(:size) { %w[Q 26] }
@@ -283,7 +255,7 @@ RSpec.describe AtCoderFriends::Generator::RubyBuiltin do
     subject { generator.generate(pbm) }
     let(:pbm) do
       AtCoderFriends::Problem.new('A') do |pbm|
-        pbm.formats_raw = formats
+        pbm.formats_src = formats
         pbm.constants = constants
         pbm.options.interactive = interactive
       end
@@ -297,11 +269,15 @@ RSpec.describe AtCoderFriends::Generator::RubyBuiltin do
       end
       let(:formats) do
         [
-          AtCoderFriends::Problem::InputFormat.new(:single, :number, %w[N]),
+          AtCoderFriends::Problem::InputFormat.new(
+            :single, :number, %w[N], []
+          ),
           AtCoderFriends::Problem::InputFormat.new(
             :varray, :number, %w[x y], %w[N]
           ),
-          AtCoderFriends::Problem::InputFormat.new(:single, :string, %w[Q]),
+          AtCoderFriends::Problem::InputFormat.new(
+            :single, :string, %w[Q], []
+          ),
           AtCoderFriends::Problem::InputFormat.new(
             :harray, :string, %w[a], %w[Q]
           )
@@ -345,7 +321,9 @@ RSpec.describe AtCoderFriends::Generator::RubyBuiltin do
       end
       let(:formats) do
         [
-          AtCoderFriends::Problem::InputFormat.new(:single, :number, %w[N Q])
+          AtCoderFriends::Problem::InputFormat.new(
+            :single, :number, %w[N Q], []
+          )
         ]
       end
       let(:constants) do